+++ NEWS.md
@@ -6,6 +6,10 @@ Bugs:
 
 * Typst footnotes with cell coordinates. Thanks to @sverrirarnors for report #456.
 
+Docs improvement:
+
+* Quarto issue with `-` as list. Thanks to Chris Goodman.
+
 ## 0.8.0
 
 New: