+++ NEWS.md
@@ -1,6 +1,6 @@
 # News
 
-## Development
+## 0.9.0
 
 New: